Fix crash in MediaQualityAnalyzer
The current MediaQualityAnalyzer is using direct function call from the multiple node module running in different thread. It causes the vilolated memory access crash when the data list is not synchronized. I modify MediaQualityAnalzyer the collecting statistics and process message in the same thread using the message queue to prevent the crash.
Modify MediaQuality member valuable to use stack memory

Bug: 262843907
Bug: 263464083
Bug: 263442080
Test: Voice Call test with device, atest ImsMediaNativeTests
Change-Id: Ifd9be2102a7b959c404674f515486f661ca932f2