Fix the infinite loop error in AudioJitterBuffer
There is an issue when the AudioJitterBuffer invoke resync method if the time difference is not fix to any statement in the loop in condition of the audio frame is SID and the time difference lower than current jitter buffer size minus 20msec.
I fix the infinite loop problem to add escape.

Bug: 275635779
Test: Verified multiple voice call test in live network
Change-Id: I0a6a47f81444219b1fdf561fa69d080dba4ac5d1