Fix the thread blocking in libimsmedia
The processEvent function in ImsMediaEventHandler was blocking the thread that called it because it was calling the SendEvent function while holding the mutex lock. This caused the thread to wait for the mutex lock to be released before it could continue executing.

The issue is reproduced when the incoming audio packet contains CMR to process and invoke the new event to ImsMediaEventHandler.

To fix this, we call mutex unlock before calling processEvent in ImsMediaEventHandler. So it can call SendEvent without blocking the main thread.

Bug: 281012954
Test: Verified the Voice/Video Call in live network, atest ImsMediaNativeTests --test-filter="*AudioManagerTest*" --rerun-until-failure 100

Change-Id: I0d95f5aae9d4135f3e997b4dec7c02c34b6a2639
Merged-In: I0d95f5aae9d4135f3e997b4dec7c02c34b6a2639
(cherry picked from commit 6610cffebd4242e7459d5d983d54ab64a17ff6ab)