The packet loss calculation error when call resumed from remote hold is resolved
When the sequence number of audio packet is reduced to backward suddenly, the AudioJitterBuffer calculates the difference to the packet lost and it will trigger the session handover to LTE. I change the code to ignore the difference over the maximum threshold 3000 which is the bigger number of possible value based on the current requirements. After applying the patch the unexpected handover is gone and packet loss rate goes to normal.

Bug: 286814418
Test: Verified the case in live network, atest ImsMediaNativeTests.
Change-Id: Idb99cced37d0ed3286feec85d33bc7b2e105644b